refactor: legacy 폴더 생성 및 import 경로 수정
